Refactor to use `createTemplateFactory`
Refactors this plugin to use `createTemplateFactory`, which is the
proper import for defining a template factory in Ember. This will
continue to be transformed by the Ember modules polyfill API, which
should be included after this plugin (and is by `ember-cli-babel`).
A test for that plugin sequencing has been added as well.

Also updated the code to properly register new bindings on the scope.
Previously, new bindings were not found, and so they were not properly
replaced/renamed by subsequent plugin passes.