Improve parallelization in CI
Previously, this wasn't even *starting* the compatibility tests until
we passed on Windows. Given that Windows is a nightmare of slowness and
the more so with GH Actions' runners... let's not. Put it in a separate
job from the main test run, and add in a macOS test to go with it for
good measure.

Along the same lines, start by just doing a basic test on ubuntu with
the lowest supported Node version, and trigger all the other tests to
kick off in parallel after that: compat tests, floating dependency
tests, and tests for other supported Node versions.