Use original name in TemplateLiteralMatch
The initial implementation would use an aliased name as the value for
`importIdentifier` in the resulting `TemplateLiteralMatch`, but it
should be the original non-alaised `importIdentifier` (e.g. it should be
the thing that matches the `StaticImportConfig` value that caused it to
be a match to begin with).

Additionally, this changes to _only_ return template literals that were
actually matched the the provided `StaticImportConfig`. Previously, if
no template literal configuration was provided we would return **all**
template literals that were found (in this case, the `importIdentifier`
and `importPath` would be `undefined` values).