Stop trying to tokenize regexes to avoid confusing bug with division
The *right* way for this addon to work is via babel or something with a
proper grammar and parser. However, our current implementation relies on
tokenization via regular expression, and the attempt to tokenize regular
expressions in code causes bugs in code that does division with a slash
character. This commit removes the "/" from the tokenization and trades
the problem of division before a template tag breaking template parsing
for the much less likely problem of improperly detecting a template tag
inside a regex.

Here's an example of some code with a template that will now fail to parse:

```gjs
const myregex = /<template>/;
<template>Hello!</template>
```

And here is what previously failed but will work after this commit:

```
const divide = () => 4 / 2;
<template>Hello!</template>
```